Solving errors and refactor
