+++ server/lib/python/cartodb_services/cartodb_services/mapbox/__init__.py
@@ -2,4 +2,3 @@ from routing import MapboxRouting, MapboxRoutingResponse
 from geocoder import MapboxGeocoder
 from isolines import MapboxIsolines, MapboxIsochronesResponse
 from matrix_client import MapboxMatrixClient
-from exceptions import ServiceException
