+++ server/lib/python/cartodb_services/cartodb_services/mapzen/isochrones.py
@@ -3,7 +3,9 @@ import json
 import re
 
 from requests.adapters import HTTPAdapter
-from exceptions import WrongParams, MalformedResult, ServiceException
+from cartodb_services.tools.exceptions import (WrongParams,
+                                               MalformedResult,
+                                               ServiceException)
 from cartodb_services.tools.qps import qps_retry
