+++ server/lib/python/cartodb_services/cartodb_services/mapzen/matrix_client.py
@@ -1,7 +1,7 @@
 import requests
 import json
 from cartodb_services.tools.qps import qps_retry
-from exceptions import ServiceException
+from cartodb_services.tools.exceptions import ServiceException
 from cartodb_services.metrics import Traceable
