+++ server/lib/python/cartodb_services/cartodb_services/tools/__init__.py
@@ -4,4 +4,4 @@ from polyline import PolyLine
 from log import Logger, LoggerConfig
 from rate_limiter import RateLimiter
 from service_manager import ServiceManager, RateLimitExceeded
-from legacy_service_manager import  LegacyServiceManager
+from legacy_service_manager import LegacyServiceManager
