+++ server/lib/python/cartodb_services/test/test_mapboxmatrix.py
@@ -2,7 +2,7 @@ import unittest
 from mock import Mock
 from cartodb_services.mapbox import MapboxMatrixClient
 from cartodb_services.mapbox.matrix_client import DEFAULT_PROFILE
-from cartodb_services.mapbox import ServiceException
+from cartodb_services.tools.exceptions import ServiceException
 from cartodb_services.tools import Coordinate
 
 VALID_TOKEN = 'pk.eyJ1IjoiYWNhcmxvbiIsImEiOiJjamJuZjQ1Zjc0Ymt4Mnh0YmFrMmhtYnY4In0.gt9cw0VeKc3rM2mV5pcEmg'
