+++ server/lib/python/cartodb_services/test/test_mapzengeocoder.py
@@ -6,7 +6,8 @@ import requests_mock
 from mock import Mock
 
 from cartodb_services.mapzen import MapzenGeocoder
-from cartodb_services.mapzen.exceptions import MalformedResult, TimeoutException
+from cartodb_services.tools.exceptions import (MalformedResult,
+                                               TimeoutException)
 
 requests_mock.Mocker.TEST_PREFIX = 'test_'
